requirements.txt
torch
torchvision

pytorch-lightning==1.6.2
torchmetrics==0.8.1
lightning-flash==0.7.4

timm==0.5.4
wandb==0.12.6
torch_optimizer==0.3.0
pytorch-metric-learning==1.3.0
rich
datasets[vision]==2.2.1

pre-commit
